API/requirements.txt
fastapi
uvicorn
scikit-learn
pydantic
sqlalchemy
sqlite
python-dotenv
